feat: ajustes review
