+++ src/app/@core/utils/general_utility.ts
@@ -1,0 +1,17 @@
+function  groupBy(xs, key) {
+    let result;
+    try {
+      const grouped = xs.reduce(function (rv, x) {
+        (rv[x[key]] = rv[x[key]] || []).push({...x});
+        return rv;
+      }, {});
+      result = Object.keys(grouped).map((i) => ({ group: i, data: grouped[i] }));
+    } catch {
+      result = [];
+    }
+    return result;
+}
+
+export {
+    groupBy
+};
